Next

+++ Python3-MQTT-To-HomeAssistant.py
@@ -10,6 +10,7 @@ import glob
 import paho.mqtt.client as mqtt
 import time
 
+# This is for my custom logger, that should make it possible to write all types of erros to files (Even the python script itself)
 class CustomLogger(logging.Logger):
     def __init__(self, name, log_filename):
         super().__init__(name)
@@ -17,22 +18,23 @@ class CustomLogger(logging.Logger):
         # Set up formatter for log messages
         formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
 
-        # Set up log file handler
-        file_handler = logging.FileHandler(log_filename)
-        file_handler.setLevel(logging.INFO)
-        file_handler.setFormatter(formatter)
-        self.addHandler(file_handler)
+        # Set up log file handler only if log_filename is provided
+        if log_filename:
+            file_handler = logging.FileHandler(log_filename)
+            file_handler.setLevel(logging.INFO)
+            file_handler.setFormatter(formatter)
+            self.addHandler(file_handler)
 
-        # Set up console handler
         console_handler = logging.StreamHandler()
         console_handler.setLevel(logging.INFO)
         console_handler.setFormatter(formatter)
         self.addHandler(console_handler)
         
+# This is for my custom logger, that should make it possible to write all types of erros to files (Even the python script itself)
 def setup_logger(log_folder, log_date):
     global err_filepath  # Use the global variable
     log_filename = f"MQTT-To-HomeAssistant-Date{log_date}.log"
-    log_filepath = os.path.join(log_folder, log_filename)
+    log_filepath = os.path.join(log_folder, log_filename) if log_folder.lower() != "no" else None
 
     # Set up logger for normal output
     logger = CustomLogger("MQTT-To-HomeAssistant-Date", log_filepath)
@@ -42,7 +44,7 @@ def setup_logger(log_folder, log_date):
 
     # Set up logger for errors
     err_filename = f"MQTT-To-HomeAssistant-Date{log_date}.err"
-    err_filepath = os.path.join(log_folder, err_filename)
+    err_filepath = os.path.join(log_folder, err_filename) if log_folder.lower() != "no" else None
     error_logger = CustomLogger("MQTT-To-HomeAssistant-Date-Error", err_filepath)
 
     # Set the error logger level
@@ -50,6 +52,7 @@ def setup_logger(log_folder, log_date):
 
     return logger, error_logger
 
+# This is where we connect to the MQTT broker when everything is ready to be sent
 def mqtt_connect(client, userdata, flags, rc, mqtt_topic, mqtt_message, logger, error_logger):
     try:
         if rc == 0:
@@ -71,6 +74,7 @@ def mqtt_connect(client, userdata, flags, rc, mqtt_topic, mqtt_message, logger, 
         print_separator(logger, error_logger)
         error_logger.error('MQTT Error message: ' + error_message)
 
+# This is for finding the newest log files to attach to the mail if logs is enabled 
 def get_newest_files(log_dir, prefix):
     files = glob.glob(os.path.join(log_dir, f"{prefix}*"))
     files.sort(key=os.path.getctime, reverse=True)
@@ -106,9 +110,7 @@ def send_mail(subject, body, recipient, attachment_files=None):
     return mail_exit_code, stderr_output.decode().strip()
 
 # This is for the send mail function
-# In case one needs to be notified of errors
-#
-# FIx and make sure to make it possible to send error message even if .out file is not created yet
+# In case one needs to be notified of succes or errors
 def MailTo(logger, error_logger, recipient, subject):
     #log_folder = os.path.join(os.path.dirname(os.path.abspath(__file__)), "logs")
     
@@ -119,30 +121,37 @@ def MailTo(logger, error_logger, recipient, subject):
     # Define subject
     #subject = "Error snapshotting or cleaning up snapshots/logs - attaching logs"
 
-    # Get the latest .log and .err files
-    newest_log, newest_err = get_newest_files(log_folder, "MQTT-To-HomeAssistant-Date")
+    # Prepare the list for attachment files
     attachment_files = []
     
+    # Get the latest .log and .err files
+    if log_folder.upper() != "NO":
+        newest_log, newest_err = get_newest_files(log_folder, "MQTT-To-HomeAssistant-Date")
+
+        # Add the latest .log and .err files to the attachment list
+        if newest_log:
+            attachment_files.append(newest_log)
+        if newest_err:
+            attachment_files.append(newest_err)
+
     # Start by creating an empty body
     body = ""
     
-    # Add the latest .log and .err files to the attachment list
-    if newest_log:
-        attachment_files.append(newest_log)
-    if newest_err:
-        attachment_files.append(newest_err)
+    
+    if log_folder.upper() != "NO":
+        # Read contents of .err file if not empty
+        if os.path.exists(err_filepath) and not os.path.getsize(err_filepath) == 0:
+            with open(newest_err, 'r') as err_file:
+                err_contents = err_file.read()
+                body += "----------\n\n.err file\n" + err_contents
 
-     # Read contents of .err file if not empty
-    if os.path.exists(err_filepath) and not os.path.getsize(err_filepath) == 0:
-        with open(newest_err, 'r') as err_file:
-            err_contents = err_file.read()
-            body += "----------\n\n.err file\n" + err_contents
-
-    # Read contents of .log file
-    if os.path.isfile(newest_log):
-        with open(newest_log, 'r') as log_file:
-            log_contents = log_file.read()
-            body += "----------\n\n.log file\n" + log_contents
+        # Read contents of .log file
+        if os.path.isfile(newest_log):
+            with open(newest_log, 'r') as log_file:
+                log_contents = log_file.read()
+                body += "----------\n\n.log file\n" + log_contents
+    else:
+        body += "----------\n\n" + "Logs has beend disabled, enable if nessecary"
 
     # Send the Mail
     mail_exit_code, stderr_output = send_mail(subject, body, recipient, attachment_files)
@@ -152,6 +161,7 @@ def MailTo(logger, error_logger, recipient, subject):
     else:
         WasMailSent(logger, error_logger, mail_exit_code, stderr_output)
 
+# This is an attempt to insure the mail was send succesfully
 def WasMailSent(logger, error_logger, MailExitCode, popenstderr=None):
     if MailExitCode == 0:
         print_separator(logger)
@@ -164,6 +174,7 @@ def WasMailSent(logger, error_logger, MailExitCode, popenstderr=None):
         error_logger.error(popenstderr)
         print_separator(logger, error_logger)
 
+# This is in case one wishes to etc. shutdown the system or run another custom script after a succesfully MQTT message has been broadcasted
 def SystemAction(logger, error_logger):
     global MailOption
     global SystemOption
@@ -215,6 +226,7 @@ def main():
     global SystemOption
     global log_folder
 
+    # Setup the arguments the script need
     parser = argparse.ArgumentParser(description='Send a message to HomeAssistant with MQTT')
     parser.add_argument('-c', '--config', required=True, help='Command: Location for config file')
     
@@ -239,11 +251,7 @@ def main():
     # Check if it is for homeassistant
     Use_HomeAssistant = (config.get('MQTT-To-HomeAssistant', 'Use_HomeAssistant'))
 
-    # Origional Log creation from SnapBeforeWatchTower.py
-    #log_date = datetime.datetime.now().strftime('%Y-%m-%d_%H_%M_%S')
-    #log_folder = os.path.join(os.path.dirname(os.path.abspath(__file__)), "logs")
-
-    os.makedirs(log_folder, exist_ok=True)
+    if log_folder.upper() != "NO": os.makedirs(log_folder, exist_ok=True)
 
     # Create separate loggers for main logs and error logs
     logger, error_logger = setup_logger(log_folder, log_date)
@@ -301,12 +309,12 @@ def main():
         error_logger.exception("An error occurred:")
         print_separator(logger, error_logger)
         if not MailOption.upper() == "NO":
-            MailTo(logger, error_logger, MailOption)
+            MailTo(logger, error_logger, MailOption, "Error pls check or enable logs if nessesary")
 
     finally:
         # Check if the .err file is empty, and remove it if it is
-        if os.path.exists(err_filepath) and os.path.getsize(err_filepath) == 0:
-            os.remove(err_filepath)
+            if log_folder.upper() != "NO" and os.path.exists(err_filepath) and os.path.getsize(err_filepath) == 0:
+                os.remove(err_filepath)
 
 if __name__ == "__main__":
     main()
